fix(module): Handle X.Y.Z version format
Starting Go 1.21, the 'go' directive in go.mod files
takes the form:

    go X.Y.Z

(https://go.dev/doc/go1.21)

Given a go.mod file with such a directive,
`gimme module` fails because it turns "1.21.5" into "1.21.5.x".

This commit fixes the issue by adding the ".x" suffix
only if the version is in the form "X.Y".

As an added bonus, this will also work for pre-releases:

    go 1.22rc2